+++ src/migrations/index.ts
@@ -1,9 +1,15 @@
 import * as migration_20260316_020144 from './20260316_020144';
+import * as migration_20260321_100401 from './20260321_100401';
 
 export const migrations = [
   {
     up: migration_20260316_020144.up,
     down: migration_20260316_020144.down,
-    name: '20260316_020144'
+    name: '20260316_020144',
+  },
+  {
+    up: migration_20260321_100401.up,
+    down: migration_20260321_100401.down,
+    name: '20260321_100401'
   },
 ];